feat(rate-limitng): error handling + error message + adding routes
Ref: #1040
