+++ packages/api/src/routes/middlewares/rate-limitng.ts
@@ -4,8 +4,14 @@ import { NextFunction, Request, Response } from "express";
 import { getDB } from "../../models/db";
 import { getCxIdOrFail } from "../util";
 
+const routeMapForError: Record<RateLimitOperation, string> = {
+  patientQuery: "patient create or update",
+  documentQuery: "documeny query start",
+  consolidatedDataQuery: "consolidated data query start",
+};
+
 /**
- * Validates the rate limit for the given operation and ratelimit sliding window.
+ * Checks the CX request for the given operation and rate limit sliding window.
  */
 export function checkRateLimit(
   operation: RateLimitOperation,
@@ -14,16 +20,16 @@ export function checkRateLimit(
   return async (req: Request, _: Response, next: NextFunction): Promise<void> => {
     const cxId = getCxIdOrFail(req);
 
-    const underlimit = await checkRateLimitCore({
+    const allowed = await checkRateLimitCore({
       cxId,
       operation,
       rateLimit,
       client: getDB().doc,
     });
 
-    if (!underlimit) {
+    if (!allowed) {
       throw new TooManyRequestsError(
-        `Too many requests for ${operation} - please reduce your request rate`
+        `Too many requests for ${routeMapForError[operation]} - please reduce your request rate for this operation`
       );
     }
